Move roborock to be under vacuum subdirectory
