+++ miio/tests/test_roidmivacuum_miot.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 import pytest
 
 from miio import RoidmiVacuumMiot
-from miio.integrations.roborock.vacuumcontainers import DNDStatus
+from miio.integrations.vacuum.roborock.vacuumcontainers import DNDStatus
 from miio.roidmivacuum_miot import (
     ChargingState,
     FanSpeed,
